plugins-core/agent-career-manager/src/requirements.txt
requests
python-dotenv
playwright
beautifulsoup4
markdownify
python-docx
pyyaml
